refactor: ui 디렉토리 구조변경
