+++ index.html
@@ -14,13 +14,13 @@
           <button id="going-watch-button" class="select__button going-watch checked">
             👁 볼 영상
           </button>
-          <button id="watched-button" class="select__button watched_button">✅ 본 영상</button>
+          <button id="watched-button" class="select__button watched-button">✅ 본 영상</button>
         </div>
         <button id="search-modal-button" class="button nav__button">🔍 검색</button>
       </nav>
-      <div class="nothing-saved"></div>
+      <div class="nothing-saved-image-container"></div>
       <div class="saved-video" hidden>
-        <ul class="saved-video-list watched"></ul>
+        <ul class="saved-video-list"></ul>
       </div>
     </main>
     <div class="modal-container hide">
