+++ src/js/handlers/searchHandle.js
@@ -1,22 +1,22 @@
 import { isEndOfScroll } from '../util/scroll.js';
 import { $ } from '../util/querySelector.js';
 import YoutubeSearchAPI from '../service/YoutubeSearchAPI.js';
-import userInterface from '../ui/userInterface.js';
+import modalUI from '../ui/modal/modalUI.js';
 
 const youtubeSearchAPI = new YoutubeSearchAPI();
 
 export const handleSearch = () => {
   try {
     youtubeSearchAPI.resetSearchResults();
-    userInterface.resetVideoList();
+    modalUI.resetVideoList();
 
-    userInterface.renderSkeletonUI();
+    modalUI.renderSkeletonUI();
 
     const searchKeyword = $('#search-input-keyword').value.trim();
     const response = youtubeSearchAPI.callSearchAPI(searchKeyword);
     youtubeSearchAPI.updateSearchResults(response);
 
-    userInterface.renderSearchResult(response);
+    modalUI.renderSearchResult(response);
   } catch (error) {
     alert(error.message);
   }
@@ -25,12 +25,12 @@ export const handleSearch = () => {
 export const handleScrollSearch = e => {
   try {
     if (isEndOfScroll(e.target)) {
-      userInterface.renderSkeletonUI();
+      modalUI.renderSkeletonUI();
 
       const response = youtubeSearchAPI.callSearchAPI();
       youtubeSearchAPI.updateSearchResults(response);
 
-      userInterface.renderAdditionalSearchResult(response);
+      modalUI.renderAdditionalSearchResult(response);
     }
   } catch (error) {
     alert(error.message);
